Let the ManifestUpdater do the final push
Previously, when calling ManifestUpdater::UpdateAll() or Update, the
callback for pushing the manifest was only called for intermediate
manifests, but not for the final one, so the caller had to do the final
push, which was confusing. With this change, the ManifestUpdater invokes
the callback at the end of the update process for the final push.

Tested: bazel test --config=windows //...