[common] Fix FileWatcherTest (#37)
There is a race condition in RecreateWatchedDir where there was a
brief period between the second dir change event and when the file
watcher was actually watching again. If the file was written during
that bried period, it would be missed. The issue could be reproduced
easily by adding a sleep here:

  // The watched directory exists and its handle is valid.
  if (!first_run) {
    ++dir_recreate_count_;
    if (dir_recreated_cb_) dir_recreated_cb_();
    Util::Sleep(1);
  }

This CL waits until the watcher is watching again.