Init rc_len regression point to `len`, not `max_size`.
It's possible `max_size` is beyond the end of the input buffer, and we don't
want to risk returning a chunk end point beyond the end of the input buffer.
The `len` value is set to the minimum of the buffer length or `max_size`.

Technically this is not required because the following code always returns a
value within the range 1..`len`, but it's better to be safe and that code
could change.